# Improve performance of Yacht Dice