This can be improved further I'm sure. This is the maximally naive & pluggable & 5 minute solution

Working around the `state_is_list` thing was quite annoying especially in this respect