@NewSoupVi 2024-06-07
I also confirmed that the output is the exact same for the same seed.