@NewSoupVi 2024-06-07
Oh, just for some benchmarks:

Singleplayer is not improved by a ton, but still a bit:
1.3s -> 1.0s

5 player multiworld is improved a ton:
21s -> 9s